// projects/synergy-library/src/lib/enums/GlobalEnum.ts
export enum DataType {
    string = 'string',
    number = 'number',
    radiobutton = 'radiobutton',
    checkbox = 'checkbox',
    button = 'button',
    collapse = 'collapse'
}


export enum IconType {
    home = 'home',
    none = 'none'
}
export enum ButtonType {
    primary = 'primary',
    secondary = 'secondary',
    failed = 'failed',
    tertiary = 'tertiary',
    none = 'none'
}

export enum ButtonSize {
    big = 'big',
    normal = 'normal',
    small = 'small',
    full = 'full'
}

export enum InputSize {
    big = 'big',
    normal = 'normal',
    small = 'small',
    full = 'full'
}